app/requirements.txt
fastapi
uvicorn
gradio_client
azure-storage-blob